[feature-display-current-selected-shape-information] Migrated the API `GetRectangle()` from `RectangleDrawer` to `ShapeDrawer`.
